# Reimplement the current alignment::Aligner without breaking API