… src/alignment/pairwise/fast.rs