Faster global, local, semiglobal alignment with less space usage. Related to https://github.com/rust-bio/rust-bio/pull/354 and https://github.com/rust-bio/rust-bio/pull/355#pullrequestreview-501498580.

**All tests are unchanged and passed. No breaking changes. The API is completely unchanged.**

Old(https://github.com/rust-bio/rust-bio/commit/a1b114321a10bd168537ec3565aa96f2030308fc): 

```
test bench_aligner_wc_global     ... bench: 549,914,573 ns/iter (+/- 66,950,334)
test bench_aligner_wc_local      ... bench: 648,835,834 ns/iter (+/- 93,166,340)
test bench_aligner_wc_semiglobal ... bench: 622,027,993 ns/iter (+/- 43,517,887)
```

New:

```
test bench_aligner_wc_global     ... bench: 117,967,967 ns/iter (+/- 11,637,569)
test bench_aligner_wc_local      ... bench: 220,659,055 ns/iter (+/- 18,904,199)
test bench_aligner_wc_semiglobal ... bench: 152,071,703 ns/iter (+/- 11,597,219)
```

I didn't implement the "custom" alignment bacause it seemed not very useful. I haven't seen this type of alignment other than in SeqAn and rust-bio. Maybe this will be discussed in another issue.