@TianyiShi2001 2020-10-03
### TODO

- [ ] update docs
- [ ] deprecation warning of `with_capacity` methods