@TianyiShi2001 2020-10-03
I want to deprecate `with_capacity` because the time taken to initialize a few vectors is almost negligible when compared to the alignment computation as a whole. 

In addition, if the aligner first aligns large sequences and allocates a big `capacity`, and then aligns short sequences, there will be a significant waste of memory